# feat(server): add Prometheus metrics endpoint