--- a/requirements.txt
+++ b/requirements.txt
@@ -4,3 +4,4 @@ aiortc
 aiohttp
 toml
 twilio
+prometheus_client