--- a/server/metrics/__init__.py
+++ b/server/metrics/__init__.py
@@ -1,0 +1,2 @@
+from .prometheus_metrics import MetricsManager
+from .stream_stats import StreamStatsManager