--- a/server/utils/utils.py
+++ b/server/utils/utils.py
@@ -1,0 +1,65 @@
+"""General utility functions."""
+
+import asyncio
+import random
+import types
+import logging
+from aiohttp import web
+from typing import List, Tuple
+
+logger = logging.getLogger(__name__)
+
+
+# Original issue: https://github.com/aiortc/aioice/pull/63
+# Copied from: https://github.com/toverainc/willow-inference-server/pull/17/files
+def patch_loop_datagram(local_ports: List[int]):
+    loop = asyncio.get_event_loop()
+    if getattr(loop, "_patch_done", False):
+        return
+
+    # Monkey patch aiortc to control ephemeral ports
+    old_create_datagram_endpoint = loop.create_datagram_endpoint
+
+    async def create_datagram_endpoint(
+        self, protocol_factory, local_addr: Tuple[str, int] = None, **kwargs
+    ):
+        # if port is specified just use it
+        if local_addr and local_addr[1]:
+            return await old_create_datagram_endpoint(
+                protocol_factory, local_addr=local_addr, **kwargs
+            )
+        if local_addr is None:
+            return await old_create_datagram_endpoint(
+                protocol_factory, local_addr=None, **kwargs
+            )
+        # if port is not specified make it use our range
+        ports = list(local_ports)
+        random.shuffle(ports)
+        for port in ports:
+            try:
+                ret = await old_create_datagram_endpoint(
+                    protocol_factory, local_addr=(local_addr[0], port), **kwargs
+                )
+                logger.debug(f"create_datagram_endpoint chose port {port}")
+                return ret
+            except OSError as exc:
+                if port == ports[-1]:
+                    # this was the last port, give up
+                    raise exc
+        raise ValueError("local_ports must not be empty")
+
+    loop.create_datagram_endpoint = types.MethodType(create_datagram_endpoint, loop)
+    loop._patch_done = True
+
+
+def add_prefix_to_app_routes(app: web.Application, prefix: str):
+    """Add a prefix to all routes in the given application.
+
+    Args:
+        app: The web application whose routes will be prefixed.
+        prefix: The prefix to add to all routes.
+    """
+    prefix = prefix.rstrip("/")
+    for route in list(app.router.routes()):
+        new_path = prefix + route.resource.canonical
+        app.router.add_route(route.method, new_path, route.handler)