This pull request introduces a `/metrics` endpoint for Prometheus to scrape stream metrics, including FPS and average FPS per stream. Additionally, it adds the `--stream-id-label` argument, allowing users to optionally include the `stream-id` label in Prometheus metrics.

## How to Test
1. **Create a Prometheus configuration file (`prometheus.yml`)** with the following content:
    ```yaml
    global:
      scrape_interval: 5s  # Scrape metrics every 5 seconds

    scrape_configs:
      - job_name: "comfystream_metrics"
        metrics_path: "/metrics"
        static_configs:
          - targets: ["localhost:8889"]  # Use "host.docker.internal:8889" on Mac/Windows Docker
    ```

2. **Start a Prometheus container** with host network access:
    ```sh
    docker run --rm -p 9090:9090 --network host -v $(pwd)/prometheus.yml:/etc/prometheus/prometheus.yml prom/prometheus
    ```
3. **Verify the metrics endpoint is working** by visiting:
   - **Comfystream metrics endpoint:** [`http://localhost:8889/metrics`](http://localhost:8889/metrics)
   - **Prometheus dashboard:** [`http://localhost:9090/`](http://localhost:9090) → **Status** → **Targets** (to confirm scraping success).
4. **Verify stream id lable**: You can also start comfystream with the `--stream-id-label` argument to add the stream ID as a prometheus lable.

## Performance Evaluation
To ensure minimal impact on system performance, I ran the profiler from [#80](https://github.com/yondonfu/comfystream/pull/80) before and after this change.

**Test setup:**
- **Workflow:** [`[sd15_multi_cnet_depthanything_face.json](https://github.com/yondonfu/comfystream/blob/main/workflows/comfystream/examples/sd15_multi_cnet_depthanything_face.json)`](https://github.com/yondonfu/comfystream/blob/main/workflows/comfystream/examples/sd15_multi_cnet_depthanything_face.json)
- **Duration:** **120s**
- **Hardware:** **NVIDIA RTX 3090**

<details>
<summary><strong>Performance Report</strong></summary>

### **Baseline Performance (Before This PR)**
- **Average FPS:** **6 FPS**
- **System Utilization:**
  - **Without `py-spy`:** 
    ```bash
    AVERAGE - CPU: 331.68%, RAM: 19147.63MB, GPU: 44.55%, VRAM: 4320.00MB
    ```
  - **With `py-spy`:**
    ```bash
    AVERAGE - CPU: 349.37%, RAM: 19181.66MB, GPU: 45.38%, VRAM: 4320.00MB
    ```
  - **Profiling Visualization:**
    ![Old pyspy profile](https://github.com/user-attachments/assets/a48ff6cc-6766-4ed5-9820-01ee5ad073b9)

### **Performance After Adding Prometheus Metrics**
- **Average FPS:** **6 FPS**
- **System Utilization:**
  - **Without `py-spy`:**
    ```bash
    AVERAGE - CPU: 342.25%, RAM: 18897.88MB, GPU: 44.93%, VRAM: 4320.00MB
    ```
  - **With `py-spy`:**
    ```bash
    AVERAGE - CPU: 347.81%, RAM: 18899.74MB, GPU: 44.30%, VRAM: 4320.00MB
    ```
  - **Profiling Visualization:**
    ![New pyspy profile](https://github.com/user-attachments/assets/dafe2775-3dae-4b93-96a8-2797716b5ff0)

</details>

### Key Takeaways
✅ **No significant impact on FPS**  
✅ **Minimal CPU & RAM overhead from Prometheus integration**  
✅ **Enables real-time observability without performance degradation**